server/requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
appnope==0.1.0
attrs==19.3.0
Authlib==0.14.1
awsebcli==3.17.1
backcall==0.1.0
bcrypt==3.1.7
black==19.10b0
bleach==3.1.1
blessed==1.17.2
botocore==1.14.17
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dnspython==1.16.0
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
Flask==1.1.1
Flask-HTTPAuth==3.3.0
Flask-RESTful==0.3.8
future==0.16.0
idna==2.7
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.4
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
marshmallow==3.4.0
mistune==0.8.4
mongoengine==0.19.1
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.6.2
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==2.2.2
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.5.2
PyJWT==1.7.1
pymongo==3.10.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.2
pyzmq==19.0.0
regex==2020.1.8
requests==2.20.1
semantic-version==2.5.0
Send2Trash==1.5.0
sentry-sdk==0.14.2
six==1.11.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.24.3
virtualenv==20.0.5
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.0
